[Y26W2-225] fix(web): 스크롤에 따라서 expand 영역 collapse 되도록 수정
